+++ packages/terra-clinical-onset-picker/src/terra-dev-site/doc/clinical-onset-picker/ClinicalOnsetPicker.1.doc.jsx
@@ -12,6 +12,8 @@ import DefaultOnset from '../example/DefaultOnset';
 import DefaultOnsetSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DefaultOnset';
 import HandledOnset from '../example/HandledOnset';
 import HandledOnsetSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/HandledOnset';
+import OnsetWithLegend from '../example/OnsetWithLegend';
+import OnsetWithLegendSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/OnsetWithLegend';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -29,6 +31,11 @@ const DocPage = () => (
         title: 'Onset picker with supplied data and handlers',
         example: <HandledOnset />,
         source: HandledOnsetSrc,
+      },
+      {
+        title: 'Onset picker with Legend',
+        example: <OnsetWithLegend />,
+        source: OnsetWithLegendSrc,
       },
     ]}
     propsTables={[